+++ src/nrepl/project-types.ts
@@ -6,6 +6,7 @@ import * as pprint from '../printer';
 import { getConfig } from '../config';
 import { keywordize, unKeywordize } from '../util/string';
 import { CljsTypes, ReplConnectSequence } from './connectSequence';
+import { config } from '../calva-fmt/src/state';
 const { parseForms, parseEdn } = require('../../out/cljs-lib/cljs-lib');
 
 export const isWin = /^win/.test(process.platform);
@@ -196,7 +197,7 @@ const cliDependencies = () => {
     }
 }
 
-const cljsDependencies = () =>  {
+const cljsDependencies = () => {
     return {
         "lein-figwheel": {
             "cider/piggieback": PIGGIEBACK_VERSION()
@@ -398,12 +399,34 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
         }
     }
     const menuSelections = connectSequence.menuSelections;
+
     const launchAliases = menuSelections ? menuSelections.cljAliases : undefined;
     let aliases: string[] = [];
+    let aliasesWithMain = [];
     if (launchAliases) {
         aliases = launchAliases.map(keywordize);
     } else {
         let projectAliases = parsed && parsed.aliases != undefined ? Object.keys(parsed.aliases) : [];
+        for (let a in projectAliases) {
+            const aliasKey = unKeywordize(projectAliases[a]);
+            if (parsed && parsed.aliases) {
+                let alias = parsed.aliases[aliasKey];
+                if (alias && alias["main-opts"] != undefined) {
+                    aliasesWithMain.push(`:${projectAliases[a]}`);
+                }
+            }
+        }
+        if (aliasesWithMain.length > 0) {
+            const alertKey = "calva.jackInMainOptsWarningEnabled";
+            if (state.extensionContext.workspaceState.get(alertKey, true)) {
+                vscode.window.showInformationMessage(`The aliases [${aliasesWithMain.join(' ')}] specify :main-opts. Unless the options start an nREPL server, Jack-in will not work with such an alias selected.`, "OK, Don't show again", "OK")
+                    .then(answer => {
+                        if (answer === "OK, Don't show again") {
+                            state.extensionContext.workspaceState.update(alertKey, false);
+                        }
+                    });
+            }
+        }
         const myAliases = getConfig().myCljAliases;
         if (myAliases && myAliases.length) {
             projectAliases = [...projectAliases, ...myAliases];
@@ -416,6 +439,7 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
             });
         }
     }
+    let selectedAliasesHasMain = aliases.filter(a => aliasesWithMain.includes(a)).length > 0;
 
     const dependencies = {
         ...cliDependencies(),
@@ -424,16 +448,6 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
     };
     const useMiddleware = [...middleware, ...(cljsType ? cljsMiddleware[cljsType] : [])];
     const aliasesOption = aliases.length > 0 ? `-A${aliases.join("")}` : '';
-    let aliasHasMain: boolean = false;
-    for (let ali in aliases) {
-        const aliasKey = unKeywordize(aliases[ali]);
-        if (parsed && parsed.aliases) {
-            let alias = parsed.aliases[aliasKey];
-            aliasHasMain = alias && alias["main-opts"] != undefined;
-        }
-        if (aliasHasMain)
-            break;
-    }
     const q = isWin ? '"' : "'";
     const dQ = isWin ? '""' : '"';
     for (let dep in dependencies)
@@ -441,7 +455,7 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
 
     let args = ["-Sdeps", `${q}${"{:deps {" + out.join(',') + "}}"}${q}`];
 
-    if (aliasHasMain) {
+    if (selectedAliasesHasMain) {
         args.push(aliasesOption);
     } else {
         args.push(aliasesOption, "-m", "nrepl.cmdline", "--middleware", `"[${useMiddleware.join(' ')}]"`);